+++ src/lib/types.ts
@@ -6,7 +6,7 @@ import { WatchPhase } from "kubernetes-fluent-client/dist/fluent/types";
 import { Logger } from "pino";
 import { PeprMutateRequest } from "./mutate-request";
 import { PeprValidateRequest } from "./validate-request";
-import { ValidateActionResponse } from "./validate-types";
+import { ValidateActionResponse } from "./common";
 import { V1PolicyRule as PolicyRule } from "@kubernetes/client-node";
 
 /**
