+++ src/lib/validate-request.test.ts
@@ -3,8 +3,8 @@
 
 import { beforeEach, describe, expect, it } from "@jest/globals";
 import { KubernetesObject } from "kubernetes-fluent-client";
-import { ValidateActionResponse, AdmissionRequest } from "./validate-types";
-import { Operation } from "./mutate-types";
+import { ValidateActionResponse, AdmissionRequest } from "./common";
+import { Operation } from "./common";
 import { PeprValidateRequest } from "./validate-request";
 describe("PeprValidateRequest", () => {
   let mockRequest: AdmissionRequest<KubernetesObject>;
